% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_AUC.R
\name{Calculate_AUC_single}
\alias{Calculate_AUC_single}
\title{Function to calculate the AUC
It breaks down the curve into a series of rectangles
of height equals to the mean OD of two adjacent measurements
and width equals to the time difference between such adjacent measurements
Number of measurements
AUC cannot be calculated with < 2 measurements
Vectors of indices of adjacent measurements
Calculate vectors of heights and widths
Return sum of rectangles as AUC}
\usage{
Calculate_AUC_single(OD, Time)
}
\description{
Function to calculate the AUC
It breaks down the curve into a series of rectangles
of height equals to the mean OD of two adjacent measurements
and width equals to the time difference between such adjacent measurements
Number of measurements
AUC cannot be calculated with < 2 measurements
Vectors of indices of adjacent measurements
Calculate vectors of heights and widths
Return sum of rectangles as AUC
}
