% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_lag.R
\name{Calculate_lag}
\alias{Calculate_lag}
\title{Calculate_lag}
\usage{
Calculate_lag(curves_df, offset_control = 0.02)
}
\arguments{
\item{curves_df}{Master data frame containing adjusted OD measurements, cut-off
values and    Runs and Layout data.}

\item{offset_control}{Scaling factor used to account for small increases in
OD during stationary phase.}
}
\value{
A data frame object that adds end of lag phase time to the input
master data frame.
}
\description{
Determine the time of transition from lag phase to
exponential phase of the growth curve.
}
