% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cumul_min-max.R
\name{cumulmax}
\alias{cumulmax}
\title{For each element i  
Calculate max of the vector starting on 
position i until end of vector
Remove i and repeat}
\usage{
cumulmax(v)
}
\description{
For each element i  
Calculate max of the vector starting on 
position i until end of vector
Remove i and repeat
}
