% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cumul_min-max.R
\name{cumulmin}
\alias{cumulmin}
\title{Calculate length of input vector
For each element i  
Calculate min of the vector starting on 
position i until end of vector
Remove i and repeat}
\usage{
cumulmin(v)
}
\description{
Calculate length of input vector
For each element i  
Calculate min of the vector starting on 
position i until end of vector
Remove i and repeat
}
