% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detect_MIC.R
\name{Detect_MIC_single}
\alias{Detect_MIC_single}
\title{Detect_MIC_single}
\usage{
Detect_MIC_single(curves_df, threshold, concentration_var, group_var)
}
\arguments{
\item{curves_df}{Master data frame containing raw OD measurements and
Runs and Layout data.}

\item{threshold}{Area under the curve (AUC) below which a bacterial strain
is considered to be inhibited by a treatment.}

\item{concentration_var}{Variable in curves_df containing concentration data}

\item{group_var}{variables used to group curves}
}
\value{
A data frame object with a variable stating whether the MIC of a given
treatment is below ("<"), equal to ("=") or greater than(">") the tested concentration
}
\description{
Calculate minimum inhibitory concentration of a single bacterial
growth curve
}
