% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detect_outliers.R
\name{Detect_outliers}
\alias{Detect_outliers}
\title{Detect_outliers}
\usage{
Detect_outliers(
  curves_df,
  gesd_alpha = 0.1,
  gesd_max_anoms = 0.2,
  iqr_alpha = 0.1,
  iqr_max_anoms = 0.2,
  group_var = "RRPPRCC"
)
}
\arguments{
\item{curves_df}{Master data frame containing OD measurements and
Runs and Layout data.}

\item{gesd_alpha}{PENDING}

\item{gesd_max_anoms}{PENDING}

\item{iqr_alpha}{PENDING}

\item{iqr_max_anoms}{PENDING}

\item{group_var}{Variables in curves_df used to group curves.}
}
\value{
PENDING
}
\description{
Detect outlier OD measurements based on residuals of spline fit of
of a series of bacterial growth curves.
}
