% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detect_outliers.R
\name{Detect_outliers_single}
\alias{Detect_outliers_single}
\title{Detect_outliers_single}
\usage{
Detect_outliers_single(
  curves_df,
  gesd_alpha,
  gesd_max_anoms,
  iqr_alpha,
  iqr_max_anoms
)
}
\arguments{
\item{curves_df}{Master data frame containing OD measurements and
Runs and Layout data.}

\item{gesd_alpha}{PENDING}

\item{gesd_max_anoms}{PENDING}

\item{iqr_alpha}{PENDING}

\item{iqr_max_anoms}{PENDING}
}
\value{
PENDING
}
\description{
Detect outlier OD measurements based on residuals of spline fit of
a single growth curve
}
