% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fix_cutoff.R
\name{Fix_cutoff}
\alias{Fix_cutoff}
\title{Fix_cutoff}
\usage{
Fix_cutoff(curves_df, RRPPs_cutoff)
}
\arguments{
\item{curves_df}{Master data frame containing adjusted OD measurements, cut-off
values and Runs and Layout data. Note that the end of lag phase time should be included.}

\item{RRPPs_cutoff}{A list of vectors with pairs of values corresponding to the `RRPP` to change and
the new `cutoff_time`.}
}
\value{
A data frame object with fixed cut off values at the runs/plates specified
}
\description{
Manually replace end of lag or exponential phase cut-off values
 	after automatic method.
}
\seealso{
[Adjust_OD()] and [Calculate_lag()] for determination of end of
lag and exponential phases
}
