% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fix_table_names.R
\name{Fix_table_names}
\alias{Fix_table_names}
\title{Fix_table_names}
\usage{
Fix_table_names(col_names)
}
\arguments{
\item{col_names}{Vector of column names to repair}
}
\value{
A vector of same length as `col_names` with repaired names
}
\description{
Fix column names in Runs and Layout tables. All column names start
with uppercase, replace spaces with underscores and executes universal name
repair.
}
\examples{
colnames(mtcars)
# [1] "mpg"  "cyl"  "disp" "hp"   "drat" "wt"   "qsec" "vs"   "am"   "gear" "carb"
Fix_table_names(colnames(mtcars))
# [1] "Mpg"  "Cyl"  "Disp" "Hp"   "Drat" "Wt"   "Qsec" "Vs"   "Am"   "Gear" "Carb"
}
