% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_curve_plots.R
\name{Make_curve_plots}
\alias{Make_curve_plots}
\title{Make_curve_plots}
\usage{
Make_curve_plots(
  curves_df,
  save_plots = FALSE,
  plots_dir = NA,
  replicate_variable = NA
)
}
\arguments{
\item{curves_df}{Master data frame containing raw OD measurements and
Runs and Layout data.}

\item{save_plots}{Should plots saved as files?}

\item{plots_dir}{Path to folder where plots will be saved.}

\item{replicate_variable}{Name of column containing replicate experiment number}
}
\value{
A single ggplot2 plot of raw OD over time, faceted by run-plate-strain-biological
replicate combination.
}
\description{
Plot all technical replicates from the same run-plate-strain-biological replicate combination
}
