% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_vline_plots.R
\name{Make_cutoff_plots}
\alias{Make_cutoff_plots}
\title{Make_cutoff_plots}
\usage{
Make_cutoff_plots(
  curves_df,
  save_plots = FALSE,
  plots_dir = NA,
  vline = "cutoff",
  scales = "fixed",
  y_limit = c(0.1, 1),
  plot_name = "Control_cutoff"
)
}
\arguments{
\item{curves_df}{Master data frame containing raw OD measurements and
Runs and Layout data.}

\item{save_plots}{Should plots saved as files?}

\item{plots_dir}{Path to folder where plots will be saved.}

\item{vline}{Vertical line to be added to the plots.
One of "cutoff" (Default): end of exponential phase, "lag": end of lag
phase, or "both".}

\item{scales}{Should scales be fixed ("fixed", the default), free ("free"), or free in one dimension ("free_x", "free_y")?}

\item{y_limit}{Vector of y-axis limits}

\item{plot_name}{Name of file to be written}
}
\value{
A single ggplot2 plot of raw OD over time of control curves, faceted
by `RRPP` (Run and Plate identifier) with vertical lines indicating end
of lag or exponential phases.
}
\description{
Plot growth curves of controls and add vertical lines indicating
end of lag exponential phases calculated with `Adjust_OD()` and `Calculate_lag`.
Growth curves are faceted by `RRPP`, an identifier for Run and Plate
(e.g `110` corresponds to Run `1 `and plate `10`). Each facet in shows the curves
corresponding to the controls of a particular plate.
}
\seealso{
[Adjust_OD()] and [Calculate_lag] for calculation of end of lag and exponential phases.
}
