% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_flat_plot.R
\name{Make_flat_plot}
\alias{Make_flat_plot}
\title{Make_flat_plot}
\usage{
Make_flat_plot(
  curves_df,
  save_plots = FALSE,
  plots_dir = NA,
  replicate_variable = NA
)
}
\arguments{
\item{curves_df}{Master data frame containing OD measurements and whether a
curve was marked as flat.}

\item{save_plots}{Should plots saved as files?}

\item{plots_dir}{Path to folder where plots will be saved.}

\item{replicate_variable}{Name of column containing replicate experiment number}
}
\value{
A list of ggplot2 plots of OD over time. Each plot contains all the
curves from a given run-plate-strain-biological replicate combination, with
curves marked as flat highlighted.
}
\description{
Plot all technical replicates from the same run-plate-strain-biological
replicate combination and highlight those marked as flat.
}
