% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_splines_plot.R
\name{Make_spline_plots}
\alias{Make_spline_plots}
\title{Make_flat_plot}
\usage{
Make_spline_plots(
  curves_df,
  OD_col = "OD",
  print_plots = FALSE,
  save_plots = FALSE,
  plots_dir = NA,
  show_outliers = FALSE,
  replicate_variable = NA
)
}
\arguments{
\item{curves_df}{PENDING}

\item{OD_col}{PENDING}

\item{print_plots}{PENDING}

\item{save_plots}{PENDING}

\item{plots_dir}{PENDING}

\item{show_outliers}{PENDING}

\item{replicate_variable}{PENDING}
}
\value{
PENDING
}
\description{
PENDING
}
