% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mark_artefacts.R
\name{Mark_artefacts}
\alias{Mark_artefacts}
\title{Determine is a given time point is below than cutoff time
And truncate curves at cutoff
Calculate ODc01
ODc01 is a normalized OD value
value of 1 represents the “maximum growth” observed in the plate, where 0 is no growth
Flag abnormal curves:
Calculate mean and sd values of ODc01 of controls
Curves with `abnormal_t0` are those with a high t0 OD values the mean and
SD values of controls
Flag abnormalities by determining points that seem to be spikes:
It assumes that OD values are constantly increasing or at least not decreasing
Determine if a point is `increased`, i.e. whether the OD at a given time
is greater that the minimum OD value in the remaining section of the curve
plus an offset value, which corresponds to n times the SD of controls at 
time 0. Then the number of spikes is added up and curves with a number above
a certain threshold are marked as abnormal
Flag abnormalities according distribution of deltas:
The mark artifacts function uses the difference between two consecutive 
measurements (delta) and two consecutive deltas (delta2) to find abnormal
curves.
It assumes that that delta and delta2 follow an exponential distribution 
within a given strain. The `fitdistrplus::fitdist` function, fits an 
exponential distribution to the observed set of deltas and extracts the rate
(a.k.a lambda) from that particular fit. Using the `pexp` function, for each
delta and delta2, it calculates the probability of observing a value equals 
to or greater than it (*note* the lower.tail = FALSE) using an exponential
distribution with the calculated rate (a.k.a. lambda or `estimate`).
Finally, it marks an abnormality if, for a given curve, the smallest observed
delta and delta2 have a probability below the given thresholds. 
The value of these thresholds can be changed according to how stringent
the abnormality detection wants to be: Smaller p_delta and p_delta2 values
make the filtering more lenient, higher values are more strict
The interpretation of this detection can be briefly explained as follows: 
for a given curve, if the minimum delta and delta2 values have a high value
in comparison to the overall distribution within a given strain
then the probability of observing it will be low, therefore the curve will 
be marked as an abnormality.}
\usage{
Mark_artefacts(
  GCs,
  sum_inc = 3,
  increased_sd = 2,
  t0_sd = 3,
  p_delta = 0.001,
  p_delta2 = 0.001
)
}
\description{
Determine is a given time point is below than cutoff time
And truncate curves at cutoff
Calculate ODc01
ODc01 is a normalized OD value
value of 1 represents the “maximum growth” observed in the plate, where 0 is no growth
Flag abnormal curves:
Calculate mean and sd values of ODc01 of controls
Curves with `abnormal_t0` are those with a high t0 OD values the mean and
SD values of controls
Flag abnormalities by determining points that seem to be spikes:
It assumes that OD values are constantly increasing or at least not decreasing
Determine if a point is `increased`, i.e. whether the OD at a given time
is greater that the minimum OD value in the remaining section of the curve
plus an offset value, which corresponds to n times the SD of controls at 
time 0. Then the number of spikes is added up and curves with a number above
a certain threshold are marked as abnormal
Flag abnormalities according distribution of deltas:
The mark artifacts function uses the difference between two consecutive 
measurements (delta) and two consecutive deltas (delta2) to find abnormal
curves.
It assumes that that delta and delta2 follow an exponential distribution 
within a given strain. The `fitdistrplus::fitdist` function, fits an 
exponential distribution to the observed set of deltas and extracts the rate
(a.k.a lambda) from that particular fit. Using the `pexp` function, for each
delta and delta2, it calculates the probability of observing a value equals 
to or greater than it (*note* the lower.tail = FALSE) using an exponential
distribution with the calculated rate (a.k.a. lambda or `estimate`).
Finally, it marks an abnormality if, for a given curve, the smallest observed
delta and delta2 have a probability below the given thresholds. 
The value of these thresholds can be changed according to how stringent
the abnormality detection wants to be: Smaller p_delta and p_delta2 values
make the filtering more lenient, higher values are more strict
The interpretation of this detection can be briefly explained as follows: 
for a given curve, if the minimum delta and delta2 values have a high value
in comparison to the overall distribution within a given strain
then the probability of observing it will be low, therefore the curve will 
be marked as an abnormality.
}
