% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mark_flat.R
\name{Mark_flat}
\alias{Mark_flat}
\title{Mark_flat}
\usage{
Mark_flat(
  curves_df,
  OD_col = "ODc01",
  OD_diff_cutoff = 0.2,
  group_var = "RRPPRCC",
  last_OD = 1
)
}
\arguments{
\item{curves_df}{Master data frame containing adjusted OD measurements and
Runs and Layout data.}

\item{OD_col}{Name of column in master data frame containing
OD measurements. Can be raw or adjusted OD.}

\item{OD_diff_cutoff}{Average difference between each of four break points in the
growth curve and and the first measurement. If this value is  below a user-defined
threshold, the curve will be flagged as flat. Smaller values make the marking
more strict.}

\item{last_OD}{Minimum OD value that a curve should reach in last measurement
so that the 'flat' label is removed. Smaller values make the marking more strict.}
}
\value{
A data frame object with the result of the test of whether the curve
	is flat added to the input master data frame.
}
\description{
Determine whether each curve in a data frame is flat.
}
