% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mark_flat.R
\name{Mark_flat_single}
\alias{Mark_flat_single}
\title{Mark_flat_single}
\usage{
Mark_flat_single(curves_df, OD_col, OD_diff_cutoff, last_OD)
}
\arguments{
\item{curves_df}{Master data frame containing adjusted OD measurements and
Runs and Layout data.}

\item{OD_col}{Name of column in master data frame containing
OD measurements. Can be raw or adjusted OD.}

\item{OD_diff_cutoff}{Average difference between each of four break points in the
growth curve and and the first measurement. If this value is  below a user-defined
threshold, the curve will be flagged as flat. Smaller values make the marking
more strict.}

\item{last_OD}{Minimum OD value that a curve should reach in last measurement
so that the 'flat' label is removed. Smaller values make the marking more strict.}
}
\value{
A logical value of whether a curve is considered flat or not.
}
\description{
Determine whether a single growth curve is flat
}
