% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neckaR-package.R
\docType{package}
\name{neckaR-package}
\alias{neckaR}
\alias{neckaR-package}
\title{neckaR: Analysis Of Bacterial Growth Curves}
\description{
A package for the analysis of bacterial growth curves obtained from optical density (OD) measurements. The neckaR package aids in the identification of abnormal growth curves and the calculation of parameters such area under the curve (AUC), maximum OD, and end of lag and exponential phases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jacodela/neckar}
  \item Report bugs at \url{https://github.com/jacodela/neckaR/issues}
}

}
\author{
\strong{Maintainer}: Jacobo de la Cuesta-Zuluaga \email{jacobo.delacuesta@med.uni-tuebingen.de} (\href{https://orcid.org/0000-0002-7369-992X}{ORCID})

Authors:
\itemize{
  \item Patrick Müller \email{patrick.mueller@med.uni-tuebingen.de}
  \item Lisa Maier \email{l.maier@uni-tuebingen.de} (\href{https://orcid.org/0000-0002-6473-4762}{ORCID}) [funder]
}

}
\keyword{internal}
