% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_OD_96.R
\name{Read_OD_96}
\alias{Read_OD_96}
\title{Read_OD_96}
\usage{
Read_OD_96(filename, sheet, duration)
}
\arguments{
\item{filename}{Full path to an Excel file with OD measurements from
TECAN plate reader}

\item{sheet}{Number of sheet to be read from Excel}

\item{duration}{Maximum number of time points measured in the experiment}
}
\value{
A data frame object that contains the OD measurements from a
	single sheet in long format
}
\description{
Reads excel files from TECAN plate reader and converts the data
	into a long data frame object. It assumes that each sheet corresponds to
	a single 96-well plate. Time is converted to hours instead of days. An identifier
	field is created, where digits corresponding to thousands and tens of thousands
	indicate the plate number [1 to #plates]; hundreds indicates the row number [1 to 8];
	tens and units are the columns [1 to 12].
}
\examples{
  # PENDING
}
