% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sort_by_run.R
\name{Sort_by_run}
\alias{Sort_by_run}
\title{Sort_by_run}
\usage{
Sort_by_run(file_list)
}
\arguments{
\item{file_list}{a character vector of file names to be sorted based on the
RUN number.}
}
\value{
A character vector of file names sorted by the RUN number.
}
\description{
sorts a vector of file names based on the substring that contains
the run information in a case-insensitive manner.
}
