% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjust_OD.R
\name{Adjust_OD}
\alias{Adjust_OD}
\title{Function that adds extra columns in the master data frame prior further QC
NOTE
!!rlang::sym(var_name) converts a string to a variable name to be used in a dplyr verb
scaling of OD values
Control curves
ODc0_adj equals the the scaled ODc0 plus the maximum ODc0 of the plate/run
(ODc0[dplyr::n()]) times a scaling factor (offset_control) times the remaining
time (Time[dplyr::n()]-Time)
The scaling factor accounts for a small increase in OD during the stationary phase
Determine the cutoff time of each plate of each run
Value is the maximum adjusted OD of control of the plate
Join raw OD values with time cutoff}
\usage{
Adjust_OD(curves_df, control_factor, control_level, offset_control)
}
\description{
Function that adds extra columns in the master data frame prior further QC
NOTE
!!rlang::sym(var_name) converts a string to a variable name to be used in a dplyr verb
scaling of OD values
Control curves
ODc0_adj equals the the scaled ODc0 plus the maximum ODc0 of the plate/run
(ODc0[dplyr::n()]) times a scaling factor (offset_control) times the remaining
time (Time[dplyr::n()]-Time)
The scaling factor accounts for a small increase in OD during the stationary phase
Determine the cutoff time of each plate of each run
Value is the maximum adjusted OD of control of the plate
Join raw OD values with time cutoff
}
