% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_AUC.R
\name{Calculate_AUC}
\alias{Calculate_AUC}
\title{Recalculate ODc01 after cut-off
Shift and scale OD measurement
Before calculating AUC, the OD is recalculated by either shifting or 
re-scaling each curve individually.}
\usage{
Calculate_AUC(GCs, ...)
}
\description{
To calculate shifted OD, the minimum OD is substracted from all time points 
of the growth curve so that the minimum is zero
}
\details{
To re-scale OD,  a constant shift is substracted as above. Then the
curve is scaled to a [0 - 1] range such that a time point with an uncorrected
OD of 1 also had an OD of 1 after correction.
Calculate AUC

To calculate the AUC the shifted and re-scaled OD values are used and an
AUC is calculated for each.

Next, the value is normalized to the control wells of a given plate and run
Remove time and OD data to retain only conditions df
The final normalized AUC selected is the one that is closest to 1 from
the two areas calculated with the shifted and the re-scaled AUC
Add conditions info
Return
}
