% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detect_spikes.R
\name{Detect_spikes}
\alias{Detect_spikes}
\title{Create data frame with the ODc0 values of t0 and lag time of each curve
Join with original df to determine if a point is a potential spike:
Spike if ODc0 is higher than in lag time. Optionally, an offset can be
added to the lag OD value to modify how sensitive the detection is. 
The higher the offset value, the less sensitive the detection of spikes}
\usage{
Detect_spikes(CGs, ODs_offset = 0)
}
\description{
Create data frame with the ODc0 values of t0 and lag time of each curve
Join with original df to determine if a point is a potential spike:
Spike if ODc0 is higher than in lag time. Optionally, an offset can be
added to the lag OD value to modify how sensitive the detection is. 
The higher the offset value, the less sensitive the detection of spikes
}
