% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fix_cutoff.R
\name{Fix_cutoff}
\alias{Fix_cutoff}
\title{Fix_cutoff}
\usage{
Fix_cutoff(df, RRPPs_cutoff, fix_lag = FALSE)
}
\arguments{
\item{curves_df}{Master data frame containing adjusted OD measurements, cut-off
values and Runs and Layout data. Note that the end of lag phase time should be included.}

\item{offset_control}{Value added to the OD of end of lag phase to adjust
how sensitive the spike detection is.}
}
\value{
A data frame object that a variable marking whether a point is a potential
	spike
}
\description{
Manually replace end of lag or exponential phase cut-off values
 	after automatic method.
}
\examples{
  # PENDING
  @export
}
\seealso{
[Adjust_OD()] and [Calculate_lag()] for determination of end of
	lag and exponential phases
}
