% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fix_table_names.R
\name{Fix_table_names}
\alias{Fix_table_names}
\title{Fix_table_names}
\usage{
Fix_table_names(col_names)
}
\arguments{
\item{col_names}{Vector of column names to repair}
}
\value{
A vector of same length as `col_names` with repaired names
}
\description{
Fix column names in Runs and Layout tables. All column names start
with uppercase, replace spaces with underscores and executes universal name
repair.
}
\examples{
  # PENDING
  @export
}
