% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_master_df.R
\name{Make_master_df}
\alias{Make_master_df}
\title{Make_master_df}
\usage{
Make_master_df(
  Data_folder,
  Data_files,
  Runs_path,
  Layout_path,
  Duration = 20,
  Design_tab_col = "Design"
)
}
\arguments{
\item{Data_folder}{Path to the folder containing Excel files from TECAN plate
reader}

\item{Data_files}{Path to Excel files from TECAN plate reader.}

\item{Runs_path}{Path to Runs (a.k.a. Tab1) table}

\item{Layout_path}{Path to Layout (a.k.a. Tab2) table}

\item{Duration}{Maximum number of time points measured in the experiment}

\item{Design_tab_col}{Name of column with design information in Runs and
Layout tables}
}
\value{
A data frame object that contains the OD measurements from a
	all the plates from all runs included, combined with the Runs and Layout
	information. Columns correspond to those included in Runs and Layout tables.
	In addition, `Make_master_df()` creates a series of variables to identify
	each curve and each measurement. They are:
	* `ID`: numeric value where the thousands and tens of thousands positions
	correspond to the plate number [1 to # of plates]. Hundreds position is
	the row number [1 to 8]. Tens and units are the columns [01 to 12].
	* `RRPPRCC`: Unique curve identifier, representing
	(Run, Run, Plate, Plate, Row (in plate [1 to 8]), Column (in plate [01 to 12]))
}
\description{
Combines OD measurements with with run and layout information
	of an experiment
}
\examples{
  # PENDING
}
